Incorrect spelling. Class: zcl_abapgit_syntax_highlighter.clas.abap
Line: 141. Missing "r" in the word "refering" 
The correct spelling of the word is "referring", not "refering". The correct spelling contains an additional letter "r".
Line: 236. Incorrect spelling.
The correct spelling of the word is "discarded", not "discarted". The typo involves the substitution of the letter "t" for the correct letter "d".