Update zcl_abapgit_ui_factory.clas.abap
Incorrect spelling of the word "cacheable." 
Line: 72. The correct spelling is "cacheable, not "cachable. "The correct spelling contains an additional letter "e".
Line: 80. The correct spelling is "cacheable, not "cachable. "The correct spelling contains an additional letter "e".
Line: 88. The correct spelling is "cacheable, not "cachable. "The correct spelling contains an additional letter "e".