Incorrect spelling of the word "Content". Class: zcl_abapgit_object_common_aff.clas.testclasses.abap
Line: 117. The correct spelling is "Content", not "Conent". "The correct spelling contains an additional letter "t". 
Line: 125. The correct spelling is "Content", not "Conent". "The correct spelling contains an additional letter "t". 
Line: 137. The correct spelling is "Content", not "Conent". "The correct spelling contains an additional letter "t".
Line: 145. The correct spelling is "Content", not "Conent". "The correct spelling contains an additional letter "t".