HTTP connection check and improved error messages
The change introduces a connection check when working with online repositories. The connection is checked in the following situations:

- Creating a new online repo
- Changing the URL in remote settings of a repo
- Opening the repository view of a repo

Errors are now shown at the bottom of the current page. This allows viewing of the error longtext, which now provides details about the attempted HTTP connection (hostname, SSL ID, and proxy). It also includes a link to the corresponding abapGit documentation.

Before:

After:

Closes #6772