SICF: Include alternate service name
Symptom: 

Deserializing an SICF service leads to a diff showing a new object with a different filename compared to the repository.

Cause:

The service has an alternative name in lower case (`icfservice-icfaltnme_orig`). When deserializing the service, it is created with an upper-case alternate name.

Fix:

Included the lower-case alternative name in the serialized file in order to create the same service definition when deserializing the object.

Closes #6847