Avoid popup on exit after saving remote settings
When exiting the repository remote settings, you got a confirmation popup although the settings were just saved. In particular, this happened when changing from a pull request to a branch or tag.

The comparison of saved to unsaved values did not take the change of the head type into consideration.