Enhance logging with message id and number
The abapGit log now collects the message id and number. This way, when displaying the log, you will be able to view the longtex of the message.

In particular, message id and number are added for activation errors.

Also fixes a bug showed the wrong longtext, if the message id of an exception was longer than 2 characters.