Feature: Change Repo Package
Time for something new: 

Moving a complete repository to a new package always was a pain. You had to do several well thought out manual steps.

Now it's as easy as entering a new package name and clicking a button:


The feature is limited to repos with "prefix" folder logic and local packages. Maybe someone else can add the transport connection.