Feature: Objects locked in multiple requests
If an object contains several parts (`LIMU` objects), for example, a function group with includes and several functions, it's possible that changes land in different transports. In such case, abapGit would show only one of the transport requests which is misleading.

The change will show "MULTIPLE" instead of a transport number. Clicking on the link will display all involved transports requests and tasks in a popup.

The feature is available in the repo view as well as in staging.

Closes #2961