Refactor: Data loss check - Table comparison
I finally found some time to tackle #1284 (from 2018). This change removes the popups that ask from confirmation of potential data loss when changing a table (or structure contained in a table) from the deserialize process to the preceeding check phase.

The table comparison is now done upfront. If you select a table in the pull dialog and the changes could lead to data loss, for example if a field is removed, then a second popup will be shown. You will have to confirm overwriting the table. If not, the table remains unchanged.

Closes #1284

It should fix #6746 as well since the check does not depend on the local/remote state anymore (@larshp). 

PS: Although this introduces a new popup, the change moves it to the UI layer. This paves the way for refactoring the pull process into a HTML page. This page will show the results of the deserialize check with checkboxes. You select objects to process and confirm to trigger the import.