Fix navigation after error calling a new page
Example: When trying to display a repo that requires a login and the login fails (or is cancelled), abapGit would fail further navigations (see #7364). abapGit was already "on the target page" (event handler), but the re-render after error would show the previous page. 

Now, in case of login errors or cancelled login popups, abapGit navigates back to the original page.